requirements-dev.txt
coverage==4.2
django-debug-toolbar==1.5
django-webtest==1.7.9
factory-boy==2.7.0
nplusone==0.7.2

